update BoggleCards styling
