+++ app/boggle/page.tsx
@@ -1,10 +1,9 @@
 "use client";
 
-import { Box, Button, styled } from "@mui/material";
-import { lightGreen, indigo } from "@mui/material/colors";
+import { Box, Button } from "@mui/material";
+import { lightGreen } from "@mui/material/colors";
 import { useState } from "react";
-import BoggleGrid from "../components/BoggleGrid";
-import WordList from "../components/WordsList";
+import BoggleCards from "../components/BoggleCards";
 
 export default function Boggle() {
   const [squareGrid, setSquareGrid] = useState<string[][]>([]);
@@ -42,19 +41,11 @@ export default function Boggle() {
     "Z",
   ];
 
-  const StartButton = styled(Button)(() => ({
-    color: "#fff",
-    backgroundColor: lightGreen[400],
-    "&:hover": {
-      backgroundColor: lightGreen[600],
-    },
-  }));
-
   function generate() {
     setStartBtnText("Shuffle");
-    setWord("")
-    setWords([])
-    setPressedTiles([])
+    setWord("");
+    setWords([]);
+    setPressedTiles([]);
 
     let grid = [];
     for (let i = 0; i < 4; i++) {
@@ -81,46 +72,34 @@ export default function Boggle() {
           height: "100%",
           margin: 2,
           display: "flex",
+          justifyContent: "space-between",
         }}
       >
-        <StartButton
+        <Button
           onClick={generate}
           variant="contained"
-          sx={{ margin: "auto", padding: "auto" }}
+          sx={{
+            margin: "auto",
+            padding: "auto",
+            color: "#fff",
+            backgroundColor: lightGreen[400],
+            "&:hover": {
+              backgroundColor: lightGreen[600],
+            },
+          }}
         >
           {startBtnText}
-        </StartButton>
+        </Button>
         {squareGrid.length > 0 && (
-          <>
-            <BoggleGrid
-              squareGrid={squareGrid}
-              words={words}
-              setWords={setWords}
-              word={word}
-              setWord={setWord}
-              pressedTiles={pressedTiles}
-              setPressedTiles={setPressedTiles}
-            />
-            <Box
-              sx={{
-                bgcolor: indigo[400],
-                maxHeight: "100%",
-                height: 500,
-                width: 200,
-                margin: "auto",
-                display: "flex",
-                justifyContent: "center",
-              }}
-            >
-              <Box>
-                {words?.length > 0 && (
-                  <>
-                    <WordList words={words} />
-                  </>
-                )}
-              </Box>
-            </Box>
-          </>
+          <BoggleCards
+            squareGrid={squareGrid}
+            words={words}
+            setWords={setWords}
+            word={word}
+            setWord={setWord}
+            pressedTiles={pressedTiles}
+            setPressedTiles={setPressedTiles}
+          />
         )}
       </Box>
     </>
